requirements.txt
flask==3.0.0
flask-socketio==5.3.6
pillow==10.2.0
gunicorn==21.2.0
eventlet==0.35.2 